+++ packages/editor-common/web/src/index.js
@@ -57,6 +57,10 @@ export { default as decorateComponentWithProps } from './Utils/decorateComponent
 export { getToolbarTheme } from './Utils/getToolbarTheme';
 export { simplePubsub } from './Utils/simplePubsub';
 export { getModalStyles, getBottomToolbarModalStyles } from './Utils/getModalStyles';
+export {
+  pluginDecorationProps,
+  componentWillReceiveDecorationProps,
+} from './Utils/resizeDecorationPropHandlers';
 
 export {
   insertLinkAtCurrentSelection,