+++ packages/plugin-html/web/src/createHtmlPlugin.js
@@ -1,5 +1,9 @@
 import createToolbar from './toolbar';
-import { createBasePlugin } from 'wix-rich-content-editor-common';
+import {
+  createBasePlugin,
+  pluginDecorationProps,
+  componentWillReceiveDecorationProps,
+} from 'wix-rich-content-editor-common';
 import { Component, DEFAULTS } from './HtmlComponent';
 import { HTML_TYPE } from './types';
 
@@ -11,6 +15,10 @@ const createHtmlPlugin = (config = {}) => {
     onOverlayClick: ({ e, pubsub, componentData }) =>
       !componentData.src ? pubsub.set('onClickTrigger', { event: e, key: 'edit' }) : null,
     component: Component,
+    pluginDecorationProps: props => {
+      return pluginDecorationProps(props);
+    },
+    componentWillReceiveDecorationProps,
     settings,
     type: HTML_TYPE,
     toolbar: createToolbar({