+++ packages/plugin-image/web/src/createImagePlugin.js
@@ -1,8 +1,8 @@
 import createToolbar from './toolbar';
 import {
   createBasePlugin,
-  PLUGIN_DECORATION_PROPS,
-  PLUGIN_DECORATIONS,
+  pluginDecorationProps,
+  componentWillReceiveDecorationProps,
 } from 'wix-rich-content-editor-common';
 import { Component, DEFAULTS } from './image-component';
 import { IMAGE_TYPE, IMAGE_TYPE_LEGACY } from './types';
@@ -24,23 +24,8 @@ const createImagePlugin = (config = {}) => {
     component: Component,
     type: IMAGE_TYPE,
     legacyType: IMAGE_TYPE_LEGACY,
-    pluginDecorationProps: (props, componentData) => {
-      const resizeableProps = PLUGIN_DECORATION_PROPS[PLUGIN_DECORATIONS.RESIZEABLE](props);
-      const size = componentData.config?.size;
-      const isInlineSize = size === 'inline';
-      const { width, ...rest } = resizeableProps.style; // eslint-disable-line no-unused-vars
-      const style = isInlineSize ? resizeableProps.style : { ...rest };
-      return { ...resizeableProps, style };
-    },
-    componentWillReceiveDecorationProps: (props, nextProps, onPropsChange) => {
-      const { width } = PLUGIN_DECORATION_PROPS[PLUGIN_DECORATIONS.RESIZEABLE](props);
-      const { width: nextWidth } = PLUGIN_DECORATION_PROPS[PLUGIN_DECORATIONS.RESIZEABLE](
-        nextProps
-      );
-      if (width !== nextWidth) {
-        onPropsChange({ size: 'inline', width: nextWidth });
-      }
-    },
+    pluginDecorationProps,
+    componentWillReceiveDecorationProps,
     toolbar: createToolbar({
       helpers,
       anchorTarget,