+++ packages/plugin-video/web/src/createVideoPlugin.js
@@ -1,7 +1,11 @@
 import createToolbar from './toolbar';
 import { Component, DEFAULTS } from './video-component';
 import { VIDEO_TYPE, VIDEO_TYPE_LEGACY } from './types';
-import { createBasePlugin } from 'wix-rich-content-editor-common';
+import {
+  createBasePlugin,
+  pluginDecorationProps,
+  componentWillReceiveDecorationProps,
+} from 'wix-rich-content-editor-common';
 
 const createVideoPlugin = (config = {}) => {
   const { helpers, t, [VIDEO_TYPE]: settings = {}, isMobile, ...rest } = config;
@@ -10,6 +14,8 @@ const createVideoPlugin = (config = {}) => {
     component: Component,
     type: VIDEO_TYPE,
     legacyType: VIDEO_TYPE_LEGACY,
+    pluginDecorationProps,
+    componentWillReceiveDecorationProps,
     toolbar: createToolbar({
       helpers,
       t,